requirements.txt
asgiref==3.3.1
Django==3.1.6
django-multiselectfield==0.1.12
gunicorn==20.0.4
Pillow==8.1.0
pytz==2021.1
sqlparse==0.4.1
